+++ flutter_tracktrail_app/lib/domain/repositories/routines_repository.dart
@@ -15,5 +15,5 @@ abstract class RoutinesRepository {
   Future<Either<String, List<RoutineEntity>>> getRoutinesByUserEmail(
       String email);
   Future<Either<String, void>> deleteRoutine(int idRoutine);
-  Future<Either<String, int>> getCompletion(int routineId);
+  Future<Either<String, double>> getCompletion(int routineId);
 }
